gofile-downloader: adhere to PEP8 and use snake case for methods naming
Is with a heavy heart that I come here to say this project now follows the convention from PEP8 for naming methods, variables and functions (no more camel case functions and methods).
This way other people wanting to contribute to this project will face less of a barrier, with all the IDEs and linters that expects the code to follow the convention.

(Still... I prefer camel case lol.)